Added logs to UsersRepository

+++ server/Models/Users/UsersRepository.cs
@@ -19,6 +19,23 @@ namespace server.Models.Users
                         where b.Name == userName
                         select b;
 
+            if (_log.IsDebugEnabled)
+            {
+                var users =  query.AsEnumerable().Select(item =>
+                new UserEntry()
+                {
+                    Name = item.Name,
+                    Pass = item.Pass,
+                    PubKey = item.PubKey
+                });
+
+                _log.Debug("Get Returned:");
+                foreach(var user in users)
+                {
+                    _log.DebugFormat("Name {0}, Password {1}, PubKey:{2}", user.Name, user.Pass, user.PubKey);
+                }
+            }
+
             return query.FirstOrDefault();
         }
 
@@ -27,19 +44,31 @@ namespace server.Models.Users
             var query = from b in db.Entries
                         select b;
 
-            return query.AsEnumerable().Select(item =>
+            var users = query.AsEnumerable().Select(item =>
                 new UserEntry()
                 {
                     Name = item.Name,
                     Pass = item.Pass,
                     PubKey = item.PubKey
                 });
+
+            if (_log.IsDebugEnabled)
+            {
+                _log.Debug("GetAll Returned:");
+                foreach (var user in users)
+                {
+                    _log.DebugFormat("Name {0}, Password {1}, PubKey:{2}", user.Name, user.Pass, user.PubKey);
+                }
+            }
+
+            return users;
         }
 
         public void Add(UserEntry user)
         {
             db.Entries.Add(user);
             db.SaveChanges();
+            _log.DebugFormat("Added User Name {0}, Password {1}, PubKey:{2}", user.Name, user.Pass, user.PubKey);
         }
 
         public void ModifyKey(string userName, string newPubKey)
@@ -49,8 +78,11 @@ namespace server.Models.Users
                         select b;
 
             var user = query.FirstOrDefault();
+            var oldPubKey = user.PubKey;
             user.PubKey = newPubKey;
             db.SaveChanges();
+            _log.DebugFormat("Modified User Name {0}, Password {1}, PubKey: :{2} -> {3}", user.Name, user.Pass, oldPubKey, user.PubKey);
+
         }
 
         public void Remove(string userName)
@@ -60,14 +92,23 @@ namespace server.Models.Users
                         select b;
 
             var user = query.FirstOrDefault();
-            db.Entries.Remove(user);
-            db.SaveChanges();
+            if(user != null)
+            {
+                db.Entries.Remove(user);
+                db.SaveChanges();
+                _log.DebugFormat("Removed User Name {0}, Password {1}, PubKey:{2}", user.Name, user.Pass, user.PubKey);
+            }
+            else
+            {
+                _log.DebugFormat("User to remove not found Name {0}, Password {1}, PubKey:{2}", user.Name, user.Pass, user.PubKey);
+            }
         }
 
         public void RemoveAll()
         {
             db.Entries.RemoveRange(db.Entries);
             db.SaveChanges();
+            _log.Debug("All users removed");
         }
         public void Dispose()
         {
